HIVE-25089: Move Materialized View rebuild code to AlterMaterializedViewRebuildAnalyzer
